fix: move sessionSseClient out of generated api/ directory
