+++ ui/desktop/src/utils/sessionSseClient.ts
